# api/openapi.yaml
openapi: 3.0.3
info:
  title: Walwatch Auto-Renewal Keeper API
  description: API for managing blob auto-renewal on Walrus/Sui.
  version: 0.1.0
  contact:
    name: Walwatch
    url: https://walwatch.io
servers:
  - url: https://api.walwatch.io/api/v1
    description: Production
  - url: http://localhost:3001/api/v1
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    AdminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
  parameters:
    OrgIdHeader:
      name: X-Org-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Organization UUID for scoping requests
...
  schemas:
...
    RenewalCountResponse:
      type: object
      properties:
        totalRenewals:
          type: integer

    # ── Audit Log ──
...
    RenewalJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        blobRegistrationId:
          type: string
          format: uuid
        policyId:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
        attempt:
          type: integer
        maxAttempts:
          type: integer
        lastError:
          type: string
          nullable: true
        scheduledFor:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time

    # ── Team ──
...
paths:
...
  /analytics/renewals:
    get:
      tags: [Analytics]
      summary: Get renewal count
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: Renewal count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenewalCountResponse'
...
  /renewal-jobs:
    get:
      tags: [Renewal Jobs]
      summary: List renewal jobs
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of renewal jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  renewalJobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/RenewalJob'
  /renewal-jobs/{id}:
    get:
      tags: [Renewal Jobs]
      summary: Get renewal job
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Renewal job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenewalJob'
  /renewal-jobs/{id}/retry:
    post:
      tags: [Renewal Jobs]
      summary: Retry renewal job
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job retry enqueued
  /renewal-jobs/{id}/cancel:
    post:
      tags: [Renewal Jobs]
      summary: Cancel renewal job
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job cancelled

  # ── Cost Engine ──
...